Self-RAG Started
Code for Self-RAG has started. I am coming into contact with a weird bug in the simple rag where it wont run the stuff with the RAGAS wrapper even though it was running before. Took the RAGAS stuff out temporarily and now it is not generating certain things that it was generating before.