Reran some tests for SelfRAG & CRAG
More focused on SelfRAG here but had to do some reworking of the nodes and decision handling due to graph recursion errors within the SelfRAG process. Developed a way to evaluate this process while keeping track of questions that caused loops so that we can filter them out and have a more accurate representation of SelfRAG's performance on this dataset. Similar work was done for CRAG, but the circumstances of the process are different from SelfRAG and therefore does not require as much attention.